API

DisjunctiveProgramming.@disjunction — Macro

@disjunction(model, expr, kw_args...)

Add a disjunction described by the expression expr, which must be a Vector of LogicalVariableRefs.

@disjunction(model, ref[i=..., j=..., ...], expr, kw_args...)

Add a group of disjunction described by the expression expr parameterized by i, j, ..., which must be a Vector of LogicalVariableRefs.

In both of the above calls, a Disjunct tag can be added to create nested disjunctions.

The recognized keyword arguments in kw_args are the following:

base_name: Sets the name prefix used to generate constraint names. It corresponds to the constraint name for scalar constraints, otherwise, the constraint names are set to base_name[...] for each index ... of the axes axes.

container: Specify the container type.

exactly1: Specify a Bool whether a constraint should be added to only allow selecting one disjunct in the disjunction.

To create disjunctions without macros, see disjunction.

DisjunctiveProgramming.@disjunctions — Macro

@disjunctions(model, args...)

Adds groups of disjunctions at once, in the same fashion as the @disjunction macro.

The model must be the first argument, and multiple disjunctions can be added on multiple lines wrapped in a begin ... end block.

The macro returns a tuple containing the disjunctions that were defined.

Example

julia model = GDPModel(); @variable(model, w); @variable(model, x); @variable(model, Y[1:4], LogicalVariable); @constraint(model, [i=1:2], w == i, Disjunct(Y[i])); @constraint(model, [i=3:4], x == i, Disjunct(Y[i])); @disjunctions(model, begin [Y[1], Y[2]] [Y[3], Y[4]] end);`

DisjunctiveProgramming.InfiniteGDPModel — Function

InfiniteGDPModel(args...; kwargs...)

Creates an InfiniteOpt.InfiniteModel that is compatible with the capabiltiies provided by DisjunctiveProgramming.jl. This requires that InfiniteOpt be imported first.

Example

julia> using DisjunctiveProgramming, InfiniteOpt

julia> InfiniteGDPModel()

DisjunctiveProgramming.InfiniteLogical — Function

InfiniteLogical(prefs...)

Allows users to create infinite logical variables. This is a tag for the @variable macro that is a combination of InfiniteOpt.Infinite and DisjunctiveProgramming.Logical. This requires that InfiniteOpt be first imported.

Example

julia> using DisjunctiveProgramming, InfiniteOpt

julia> model = InfiniteGDPModel();

julia> @infinite_parameter(model, t in [0, 1]);

julia> @infinite_parameter(model, x[1:2] in [-1, 1]);

julia> @variable(model, Y, InfiniteLogical(t, x)) # creates Y(t, x) in {True, False}
Y(t, x)

DisjunctiveProgramming._free_variable_info — Method

_free_variable_info()::JuMP.VariableInfo

Creates a blank JuMP.VariableInfo object with no bounds, no fixed value, no start value, and neither binary nor integer constraints.

Returns

A JuMP.VariableInfo object with all flags set to false and all numeric values set to NaN.

DisjunctiveProgramming._make_variable_object — Method

_make_variable_object(props::VariableProperties)::Any

Constructs a JuMP variable object from the given VariableProperties. If the variable_type field is nothing, dispatches to JuMP.build_variable with only the variable info; otherwise, passes variable_type as a tag/type.

Returns a JuMP variable object that can be added to a model.

DisjunctiveProgramming.binary_variable — Method

binary_variable(
    vref::LogicalVariableRef
    )::Union{JuMP.AbstractVariableRef, JuMP.GenericAffExpr}

Returns the underlying binary variable for the logical variable vref which is used in the reformulated model. This is helpful to embed logical variables in algebraic constraints. If vref has a logical complement then an expression of the form 1 - bvref is returned where bvref is the binary variable of the logical complement variable.

DisjunctiveProgramming.collect_all_vars — Method

collect_all_vars(model::JuMP.AbstractModel)

Returns all variable references in the model. Extend this for model types that have additional ref types (e.g., derivatives).

DisjunctiveProgramming.copy_and_reformulate — Method

copy_and_reformulate(model, decision_vars, reform_method, method)

Copy the GDP model, reformulate the copy with reform_method, and wrap in a GDPSubmodel. The original model is not modified. The copy's objective is rewritten in terms of the copied variables.

DisjunctiveProgramming.copy_gdp_data — Method

copy_gdp_data(
    model::JuMP.AbstractModel,
    new_model::JuMP.AbstractModel,
    ref_map::JuMP.GenericReferenceMap
)::Dict{LogicalVariableRef, LogicalVariableRef}

Copy all GDP-specific data from model to new_model, including logical variables, logical constraints, disjunct constraints, and disjunctions. This function is called automatically by copy_gdp_model after JuMP.copy_model has copied the base model structure.

Arguments

model::JuMP.AbstractModel: The source model containing GDP data to copy.

new_model::JuMP.AbstractModel: The destination model that will receive the copied GDP data.

ref_map::JuMP.GenericReferenceMap: The reference map from JuMP.copy_model that maps old variable references to new ones.

Returns

Dict{LogicalVariableRef, LogicalVariableRef}: A mapping from old logical variable references to new logical variable references.

DisjunctiveProgramming.copy_gdp_model — Method

copy_gdp_model(model::JuMP.AbstractModel)

Create a copy of a GDPModel, including all variables, constraints, and GDP-specific data (logical variables, disjunctions, etc.).

Arguments

model::JuMP.AbstractModel: The GDP model to copy.

Returns A tuple (new_model, ref_map, lv_map) where:

new_model: The copied model.

ref_map::JuMP.GenericReferenceMap: Maps old variable and constraint references to new ones.

lv_map::Dict{LogicalVariableRef, LogicalVariableRef}: Maps old logical variable references to new ones.

Example

using DisjunctiveProgramming, HiGHS
model = GDPModel(HiGHS.Optimizer)
@variable(model, x)
@variable(model, Y[1:2], LogicalVariable)
@constraint(model, x <= 10, Disjunct(Y[1]))
@constraint(model, x >= 20, Disjunct(Y[2]))
@disjunction(model, Y)

new_model, ref_map, lv_map = copy_gdp_model(model)

DisjunctiveProgramming.copy_model_with_constraints — Method

copy_model_with_constraints(model, constraints, method)

Create a new model with only the variables (and their bounds) from model and the selected constraints. Returns a GDPSubmodel wrapping the new model.

DisjunctiveProgramming.create_variable — Method

create_variable(model::JuMP.AbstractModel, props::VariableProperties)::JuMP.AbstractVariableRef

Creates and adds a JuMP variable to model using the properties specified in props. This function applies all variable attributes from the VariableProperties object including binary/integer constraints, bounds, fixed values, start values, and constraint sets.

The function first creates a variable object using _make_variable_object, optionally applies any constraint sets if props.set is not nothing, and then adds the variable to the model with the specified name.

DisjunctiveProgramming.disaggregate_expression — Method

disaggregate_expression(
    model::JuMP.AbstractModel,
    expr,
    bvref::Union{JuMP.AbstractVariableRef, JuMP.GenericAffExpr},
    method::_Hull
)

Disaggregate an expression for the Hull reformulation. This function is dispatched based on the expression type:

vref::JuMP.AbstractVariableRef: Returns the disaggregated variable if it exists, otherwise returns the original variable (for binary variables or nested disaggregated variables).

aff::JuMP.GenericAffExpr: Disaggregates each term in the affine expression.

quad::JuMP.GenericQuadExpr: Disaggregates both the affine and quadratic parts of the expression.

The disaggregated expression is multiplied by the binary indicator variable bvref to enforce the disjunctive constraint.

DisjunctiveProgramming.disjunction — Function

disjunction(
    model::JuMP.AbstractModel,
    disjunct_indicators::Vector{LogicalVariableRef},
    [nested_tag::Disjunct],
    [name::String = ""];
    [exactly1::Bool = true]
)

Create a disjunction comprised of disjuncts with indicator variables disjunct_indicators and add it to model. For nested disjunctions, the nested_tag is required to indicate which disjunct it will be part of in the parent disjunction. By default, exactly1 adds a constraint of the form @constraint(model, disjunct_indicators in Exactly(1)) only allowing one of the disjuncts to be selected; this is required for certain reformulations like Hull. For nested disjunctions, exactly1 creates a constraint of the form @constraint(model, disjunct_indicators in Exactly(nested_tag.indicator)). To conveniently generate many disjunctions at once, see @disjunction and @disjunctions.

DisjunctiveProgramming.gdp_data — Method

gdp_data(model::JuMP.AbstractModel)::GDPData

Extract the GDPData from a GDPModel.

DisjunctiveProgramming.get_constant — Method

get_constant(expr)

Returns the constant portion of an expression. Extendable for model types where additional terms should be treated as constants.

DisjunctiveProgramming.get_variable_info — Method

get_variable_info(vref::JuMP.AbstractVariableRef; kwargs...)::JuMP.VariableInfo

Extracts variable information from a JuMP variable reference and returns a JuMP.VariableInfo object. This function retrieves bounds, fixed values, start values, and binary/integer status from the variable reference.

Keyword Arguments

has_lb::Bool: Whether the variable has a lower bound (default: queries vref)

has_ub::Bool: Whether the variable has an upper bound (default: queries vref)

has_fix::Bool: Whether the variable is fixed (default: queries vref)

has_start::Bool: Whether the variable has a start value (default: queries vref)

has_binary::Bool: Whether the variable is binary (default: queries vref)

has_integer::Bool: Whether the variable is integer (default: queries vref)

lower_bound: The lower bound value (default: queries vref if has_lb, else 0)

upper_bound: The upper bound value (default: queries vref if has_ub, else 0)

fixed_value: The fixed value (default: queries vref if has_fix, else 0)

start_value: The start value (default: queries vref if has_start, else 0)

binary::Bool: Binary status (default: queries vref)

integer::Bool: Integer status (default: queries vref)

Returns

A JuMP.VariableInfo object containing all the variable's attributes.

DisjunctiveProgramming.has_logical_complement — Method

has_logical_complement(vref::LogicalVariableRef)::Bool

Return a Bool whether a vref is a logical complement of another logical variable.

DisjunctiveProgramming.is_gdp_model — Method

is_gdp_model(model::JuMP.AbstractModel)::Bool

Return if model was created via the GDPModel constructor.

DisjunctiveProgramming.prepare_max_M_objective — Method

prepare_max_M_objective(model, obj::ScalarConstraint, sub::GDPSubmodel)

Convert a constraint into an objective expression for M-value maximization. Returns a single JuMP expression to pass to raw_M.

DisjunctiveProgramming.raw_M — Method

raw_M(sub::GDPSubmodel, objective, method::_MBM)

Maximize objective over sub to obtain one raw M value for MBM. Returns max(obj_value, 0) on optimal, nothing on infeasible (signals the constraint is redundant in the combined region), or method.default_M otherwise (unbounded, numerical failure, etc).

DisjunctiveProgramming.reformulate_and_relax — Method

reformulate_and_relax(model, decision_vars, reform_method, method)

Reformulate the model in-place with reform_method and relax integrality. Returns (GDPSubmodel, undo_fn) where undo_fn restores integrality.

DisjunctiveProgramming.reformulate_disjunct_constraint — Method

reformulate_disjunct_constraint(
    model::JuMP.AbstractModel,  
    con::JuMP.AbstractConstraint, 
    bvref::Union{JuMP.AbstractVariableRef, JuMP.GenericAffExpr},
    method::AbstractReformulationMethod
)

Extension point for reformulation method method to reformulate disjunction constraint con over each constraint. If method needs to specify how to reformulate the entire disjunction, see reformulate_disjunction.

DisjunctiveProgramming.reformulate_disjunction — Method

reformulate_disjunction(
    model::JuMP.AbstractModel, 
    disj::Disjunction,
    method::AbstractReformulationMethod
) where {T<:Disjunction}

Reformulate a disjunction using the specified method. Current reformulation methods include BigM, Hull, and Indicator. This method can be extended for other reformulation techniques.

The disj field is the ConstraintData object for the disjunction, stored in the disjunctions field of the GDPData object.

DisjunctiveProgramming.reformulate_model — Function

reformulate_model(model::JuMP.AbstractModel, method::AbstractSolutionMethod = BigM())

Reformulate a GDPModel using the specified method. Prior to reformulation, all previous reformulation variables and constraints are deleted.

DisjunctiveProgramming.relax_logical_vars — Method

relax_logical_vars(model::JuMP.AbstractModel)

Relax the binary variables associated with logical indicators to continuous variables in [0, 1]. Returns a vector of the relaxed variable references, which can be passed to unrelax_logical_vars to restore integrality.

DisjunctiveProgramming.requires_exactly1 — Method

requires_exactly1(method::AbstractReformulationMethod)

Return a Bool whether method requires that Exactly 1 disjunct be selected as true for each disjunction. For new reformulation method types, this should be extended to return true if such a constraint is required (defaults to false otherwise).

DisjunctiveProgramming.requires_variable_bound_info — Method

requires_variable_bound_info(method::AbstractReformulationMethod)::Bool

Return a Bool whether method requires variable bound information accessed via variable_bound_info. This should be extended for new AbstractReformulationMethod methods if needed (defaults to false). If a new method does require variable bound information, then set_variable_bound_info should also be extended.

DisjunctiveProgramming.set_variable_bound_info — Function

set_variable_bound_info(vref, method::AbstractReformulationMethod)::Tuple{<:Number, <:Number}

Returns a tuple of the form (lower_bound, upper_bound) which are the bound information needed by method to reformulate disjunctions. This only needs to be implemented for methods where requires_variable_bound_info(method) = true. These bounds can later be accessed via variable_bound_info.

DisjunctiveProgramming.unrelax_logical_vars — Method

unrelax_logical_vars(
    binary_refs::Vector{<:JuMP.AbstractVariableRef}
    )

Restore the binary constraint on variables previously relaxed by relax_logical_vars.

DisjunctiveProgramming.variable_bound_info — Method

variable_bound_info(vref::JuMP.AbstractVariableRef)::Tuple{<:Number, <:Number}

Returns a tuple of the form (lower_bound, upper_bound) needed to implement reformulation methods. Only works if requires_variable_bound_info is implemented.

DisjunctiveProgramming.variable_copy — Method

variable_copy(model::JuMP.AbstractModel, vref::JuMP.AbstractVariableRef)::JuMP.AbstractVariableRef

Creates a copy of the variable vref in the given model. The new variable will have the same properties (bounds, fixed value, start value, name, and type) as vref but will be added to model as a distinct variable. This is useful for transferring variables between models or duplicating variable definitions in reformulations.

JuMP.add_constraint — Function

JuMP.add_constraint(
    model::JuMP.AbstractModel,
    con::_DisjunctConstraint,
    name::String = ""
)::DisjunctConstraintRef

Extend JuMP.add_constraint to add a Disjunct to a GDPModel. The constraint is added to the GDPData in the .ext dictionary of the GDPModel.

JuMP.add_constraint — Method

function JuMP.add_constraint(
    model::JuMP.GenericModel,
    c::JuMP.ScalarConstraint{_LogicalExpr, MOI.EqualTo{Bool}},
    name::String = ""
)

Extend JuMP.add_constraint to allow creating logical proposition constraints for a GDPModel with the @constraint macro. Users should define logical constraints via the syntax @constraint(model, logical_expr := true).

JuMP.add_constraint — Method

function JuMP.add_constraint(
    model::JuMP.GenericModel,
    c::VectorConstraint{<:F, S},
    name::String = ""
) where {F <: Vector{<:LogicalVariableRef}, S <: AbstractCardinalitySet}

Extend JuMP.add_constraint to allow creating logical cardinality constraints for a GDPModel with the @constraint macro.

JuMP.add_variable — Function

JuMP.add_variable(
    model::JuMP.Model,
    v::LogicalVariable, 
    name::String = ""
    )::LogicalVariableRef

Extend JuMP.add_variable for LogicalVariables. This helps enable @variable(model, [var_expr], Logical).

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::_MOI.AbstractScalarSet,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add constraints to disjuncts. This in combination with JuMP.add_constraint enables the use of @constraint(model, [name], constr_expr, tag), where tag is a Disjunct(::Type{LogicalVariableRef}). The user must specify the LogicalVariable to use as the indicator for the _DisjunctConstraint being created.

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::MathOptInterface.Nonnegatives,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add VectorConstraints to disjuncts.

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::MathOptInterface.Nonpositives,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add VectorConstraints to disjuncts.

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::MathOptInterface.Zeros,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add VectorConstraints to disjuncts.

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::Nonnegatives,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add VectorConstraints to disjuncts.

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::Nonpositives,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add VectorConstraints to disjuncts.

JuMP.build_constraint — Method

JuMP.build_constraint(
    _error::Function,
    func,
    set::Zeros,
    tag::Disjunct
)::_DisjunctConstraint

Extend JuMP.build_constraint to add VectorConstraints to disjuncts.

JuMP.build_constraint — Method

function JuMP.build_constraint(
    _error::Function,
    func::AbstractVector{T},
    set::S
) where {T <: LogicalVariableRef, S <: Union{Exactly, AtLeast, AtMost}}

Extend JuMP.build_constraint to add logical cardinality constraints to a GDPModel. This in combination with JuMP.add_constraint enables the use of @constraint(model, [name], logical_expr in set), where set can be either of the following cardinality sets: AtLeast(n), AtMost(n), or Exactly(n).

Example

To select exactly 1 logical variable Y to be true, do (the same can be done with AtLeast(n) and AtMost(n)):

using DisjunctiveProgramming
model = GDPModel();
@variable(model, Y[i = 1:2], LogicalVariable);
@constraint(model, [Y[1], Y[2]] in Exactly(1));

JuMP.build_variable — Method

JuMP.build_variable(
    _error::Function,
    info::JuMP.VariableInfo, 
    ::Union{Type{Logical}, Logical};
    [logical_complement::Union{Nothing, LogicalVariableRef} = nothing]
    )::LogicalVariable

Extend JuMP.build_variable to work with logical variables. This in combination with JuMP.add_variable enables the use of @variable(model, [var_expr], Logical). Optionally, a logical_complement can be given which provides a logical variable that is the complement of this one (common for disjunctions with two disjuncts).

JuMP.constraint_object — Method

JuMP.constraint_object(cref::DisjunctConstraintRef)

Return the underlying constraint data for the constraint referenced by cref.

JuMP.constraint_object — Method

JuMP.constraint_object(cref::DisjunctionRef)

Return the underlying constraint data for the constraint referenced by cref.

JuMP.constraint_object — Method

JuMP.constraint_object(cref::LogicalConstraintRef)

Return the underlying constraint data for the constraint referenced by cref.

JuMP.copy_extension_data — Method

JuMP.copy_extension_data(
    data::GDPData,
    new_model::JuMP.AbstractModel,
    old_model::JuMP.AbstractModel
)::GDPData

Extend JuMP.copy_extension_data to initialize an empty GDPData object for the copied model. This is the first step in the model copying process and is automatically called by JuMP.copy_model. The actual GDP data (logical variables, disjunctions, etc.) is copied separately via copy_gdp_data.

JuMP.delete — Method

JuMP.delete(model::JuMP.AbstractModel, cref::DisjunctConstraintRef)

Delete a disjunct constraint from the GDP model.

JuMP.delete — Method

JuMP.delete(model::JuMP.AbstractModel, cref::DisjunctionRef)

Delete a disjunction constraint from the GDP model.

JuMP.delete — Method

JuMP.delete(model::JuMP.AbstractModel, cref::LogicalConstraintRef)

Delete a logical constraint from the GDP model.

JuMP.delete — Method

JuMP.delete(model::JuMP.AbstractModel, vref::LogicalVariableRef)::Nothing

Delete the logical variable associated with vref from the GDP model.

JuMP.fix — Method

JuMP.fix(vref::LogicalVariableRef, value::Bool)::Nothing

Fix a logical variable to a value. Update the fixing constraint if one exists, otherwise create a new one.

JuMP.fix_value — Method

JuMP.fix_value(vref::LogicalVariableRef)::Bool

Return the value to which a logical variable is fixed.

JuMP.index — Method

JuMP.index(cref::DisjunctConstraintRef)

Return the index constraint associated with cref.

JuMP.index — Method

JuMP.index(cref::DisjunctionRef)

Return the index constraint associated with cref.

JuMP.index — Method

JuMP.index(cref::LogicalConstraintRef)

Return the index constraint associated with cref.

JuMP.index — Method

JuMP.index(vref::LogicalVariableRef)::LogicalVariableIndex

Return the index of logical variable that associated with vref.

JuMP.is_fixed — Method

JuMP.is_fixed(vref::LogicalVariableRef)::Bool

Return true if vref is a fixed variable. If true, the fixed value can be queried with fix_value.

JuMP.is_valid — Method

JuMP.is_valid(model::JuMP.AbstractModel, cref::DisjunctConstraintRef)

Return true if cref refers to a valid constraint in the GDP model.

JuMP.is_valid — Method

JuMP.is_valid(model::JuMP.AbstractModel, cref::DisjunctionRef)

Return true if cref refers to a valid constraint in the GDP model.

JuMP.is_valid — Method

JuMP.is_valid(model::JuMP.AbstractModel, cref::LogicalConstraintRef)

Return true if cref refers to a valid constraint in the GDP model.

JuMP.is_valid — Method

JuMP.is_valid(model::JuMP.AbstractModel, vref::LogicalVariableRef)::Bool

Return true if vref refers to a valid logical variable in GDP model.

JuMP.isequal_canonical — Method

JuMP.isequal_canonical(v::LogicalVariableRef, w::LogicalVariableRef)::Bool

Return true if v and w refer to the same logical variable in the same GDP model.

JuMP.name — Method

JuMP.name(cref::DisjunctConstraintRef)

Get a constraint's name attribute.

JuMP.name — Method

JuMP.name(cref::DisjunctionRef)

Get a constraint's name attribute.

JuMP.name — Method

JuMP.name(cref::LogicalConstraintRef)

Get a constraint's name attribute.

JuMP.name — Method

JuMP.name(vref::LogicalVariableRef)::String

Get a logical variable's name attribute.

JuMP.owner_model — Method

JuMP.owner_model(cref::DisjunctConstraintRef)

Return the model to which cref belongs.

JuMP.owner_model — Method

JuMP.owner_model(cref::DisjunctionRef)

Return the model to which cref belongs.

JuMP.owner_model — Method

JuMP.owner_model(cref::LogicalConstraintRef)

Return the model to which cref belongs.

JuMP.owner_model — Method

JuMP.owner_model(vref::LogicalVariableRef)::JuMP.AbstractModel

Return the GDP model to which vref belongs.

JuMP.set_name — Method

JuMP.set_name(cref::DisjunctConstraintRef, name::String)

Set a constraint's name attribute.

JuMP.set_name — Method

JuMP.set_name(cref::DisjunctionRef, name::String)

Set a constraint's name attribute.

JuMP.set_name — Method

JuMP.set_name(cref::LogicalConstraintRef, name::String)

Set a constraint's name attribute.

JuMP.set_name — Method

JuMP.set_name(vref::LogicalVariableRef, name::String)::Nothing

Set a logical variable's name attribute.

JuMP.set_start_value — Method

JuMP.set_start_value(vref::LogicalVariableRef, value::Union{Nothing, Bool})::Nothing

Set the start value of the logical variable vref.

Pass nothing to unset the start value.

JuMP.start_value — Method

JuMP.start_value(vref::LogicalVariableRef)::Bool

Return the start value of the logical variable vref.

JuMP.unfix — Method

JuMP.unfix(vref::LogicalVariableRef)::Nothing

Delete the fixed value of a logical variable.

JuMP.value — Method

JuMP.value(vref::LogicalVariableRef)::Bool

Returns the optimized value of vref. This dispatches on value(binary_variable(vref)) and then rounds to the closest Bool value.

DisjunctiveProgramming.AbstractCardinalitySet — Type

AbstractCardinalitySet <: MOI.AbstractVectorSet

An abstract type for cardinality sets _MOIAtLeast, _MOIExactly, and _MOIAtMost.

DisjunctiveProgramming.AbstractReformulationMethod — Type

AbstractReformulationMethod <: AbstractSolutionMethod

An abstract type for reformulation approaches used to solve GDPModels.

DisjunctiveProgramming.AbstractSolutionMethod — Type

AbstractSolutionMethod

An abstract type for solution methods used to solve GDPModels.

DisjunctiveProgramming.AtLeast — Type

AtLeast{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet

Convenient alias for using _MOIAtLeast.

DisjunctiveProgramming.AtMost — Type

AtMost{T<:Union{Int,LogicalVariableRef}} <: JuMP.AbstractVectorSet

Convenient alias for using _MOIAtMost.

DisjunctiveProgramming.BigM — Type

BigM{T} <: AbstractReformulationMethod

A type for using the big-M reformulation approach for disjunctive constraints.

Fields

value::T: Big-M value (default = 1e9).

tight::Bool: Attempt to tighten the Big-M value (default = true)?

DisjunctiveProgramming.ConstraintData — Type

ConstraintData{C <: JuMP.AbstractConstraint}

A type for storing constraint objects in GDPData and any meta-data they possess.

Fields

constraint::C: The constraint.

name::String: The name of the proposition.

DisjunctiveProgramming.CuttingPlanes — Type

CuttingPlanes{O,T} <: AbstractReformulationMethod

A type for using the cutting planes approach for disjunctive constraints.

Fields

optimizer::O: Optimizer to use when solving mini-models (required).

max_iter::Int: Number of iterations (default = 3).

seperation_tolerance::T: Tolerance for the separation problem (default = 1e-6).

final_reform_method::AbstractReformulationMethod: Final reformulation

method to use after cutting planes (default = BigM()).

M_value::T: Big-M value to use in the final reformulation (default = 1e9).

DisjunctiveProgramming.Disjunct — Type

Disjunct

Used as a tag for constraints that will be used in disjunctions. This is done via the following syntax:

julia> @constraint(model, [constr_expr], Disjunct)

julia> @constraint(model, [constr_expr], Disjunct(lvref))

where lvref is a LogicalVariableRef that will ultimately be associated with the disjunct the constraint is added to. If no lvref is given, then one is generated when the disjunction is created.

DisjunctiveProgramming.DisjunctConstraintIndex — Type

DisjunctConstraintIndex

A type for storing the index of a Disjunct.

Fields

value::Int64: The index value.

DisjunctiveProgramming.DisjunctConstraintRef — Type

DisjunctConstraintRef{M <: JuMP.AbstractModel}

A type for looking up disjunctive constraints.

DisjunctiveProgramming.Disjunction — Type

Disjunction{M <: JuMP.AbstractModel} <: JuMP.AbstractConstraint

A type for a disjunctive constraint that is comprised of a collection of disjuncts of indicated by a unique LogicalVariableIndex.

Fields

indicators::Vector{LogicalVariableref}: The references to the logical variables

(indicators) that uniquely identify each disjunct in the disjunction.

nested::Bool: Is this disjunction nested within another disjunction?

DisjunctiveProgramming.DisjunctionIndex — Type

DisjunctionIndex

A type for storing the index of a Disjunction.

Fields

value::Int64: The index value.

DisjunctiveProgramming.DisjunctionRef — Type

DisjunctionRef{M <: JuMP.AbstractModel}

A type for looking up disjunctive constraints.

DisjunctiveProgramming.Exactly — Type

Exactly <: JuMP.AbstractVectorSet

Convenient alias for using _MOIExactly.

DisjunctiveProgramming.GDPData — Type

GDPData{M <: JuMP.AbstractModel, V <: JuMP.AbstractVariableRef, CrefType, ValueType}

The core type for storing information in a GDPModel.

DisjunctiveProgramming.GDPModel — Method

GDPModel([optimizer]; [kwargs...])::JuMP.Model

GDPModel{T}([optimizer]; [kwargs...])::JuMP.GenericModel{T}

GDPModel{M <: JuMP.AbstractModel, VrefType, CrefType}([optimizer], [args...]; [kwargs...])::M

The core model object for building general disjunction programming models.

DisjunctiveProgramming.GDPSubmodel — Type

GDPSubmodel{M, V, W}

A unified submodel wrapper used by MBM and cutting plane reformulations. It encapsulates a flat JuMP optimization submodel built from a single disjunct's feasible region, along with mappings back to the original model's variables.

Fields

model::M: The JuMP submodel representing a disjunct's feasible region (constraints and variable bounds).

decision_vars::Vector{V}: Ordered decision variables in the submodel, matching the original model's ordering.

fwd_map::Dict{V, Vector{W}}: Forward map from original model variables to their submodel counterparts.

DisjunctiveProgramming.Hull — Type

Hull{T} <: AbstractReformulationMethod

A type for using the convex hull reformulation approach for disjunctive constraints.

Fields

value::T: epsilon value for nonlinear hull reformulations (default = 1e-6).

DisjunctiveProgramming.Indicator — Type

Indicator <: AbstractReformulationMethod

A type for using indicator constraint approach for linear disjunctive constraints.

DisjunctiveProgramming.Logical — Type

Logical{T}

Tag for creating logical variables using @variable. Most often this will be used to enable the syntax:

@variable(model, var_expr, Logical, [kwargs...])

which creates a LogicalVariable that will ultimately be reformulated into a binary variable of the form:

@variable(model, var_expr, Bin, [kwargs...])

To include a tag that is used to create the reformulated variables, the syntax becomes:

@variable(model, var_expr, Logical(MyTag()), [kwargs...])

which creates a LogicalVariable that is associated with MyTag() such that the reformulation binary variables are of the form:

@variable(model, var_expr, Bin, MyTag(), [kwargs...])

DisjunctiveProgramming.LogicalConstraintIndex — Type

LogicalConstraintIndex

A type for storing the index of a logical constraint.

Fields

value::Int64: The index value.

DisjunctiveProgramming.LogicalConstraintRef — Type

LogicalConstraintRef{M <: JuMP.AbstractModel}

A type for looking up logical constraints.

DisjunctiveProgramming.LogicalVariable — Type

LogicalVariable <: JuMP.AbstractVariable

A variable type the logical variables associated with disjuncts in a Disjunction.

Fields

fix_value::Union{Nothing, Bool}: A fixed boolean value if there is one.

start_value::Union{Nothing, Bool}: An initial guess if there is one.

logical_complement::Union{Nothing, LogicalVariableRef}: The logical complement of this variable if there is one.

DisjunctiveProgramming.LogicalVariableData — Type

LogicalVariableData

A type for storing LogicalVariables and any meta-data they possess.

Fields

variable::LogicalVariable: The logical variable object.

name::String: The name of the variable.

DisjunctiveProgramming.LogicalVariableIndex — Type

LogicalVariableIndex

A type for storing the index of a LogicalVariable.

Fields

value::Int64: The index value.

DisjunctiveProgramming.LogicalVariableRef — Type

LogicalVariableRef{M <: JuMP.AbstractModel}

A type for looking up logical variables.

DisjunctiveProgramming.MBM — Type

MBM{O, T, L <: LogicalVariableRef} <: AbstractReformulationMethod

A type for using the multiple big-M reformulation approach for disjunctive constraints.

Fields

optimizer::O: Optimizer to use when solving mini-models (required).

default_M::T: Default big-M value to use if no big-M is specified for a logical variable (1e9).

DisjunctiveProgramming.PSplit — Type

PSplit <: AbstractReformulationMethod

A type for using the P-split reformulation approach for disjunctive constraints. This method partitions variables into groups and handles each group separately.

Constructors

PSplit(partition::Vector{Vector{V}}): Create a PSplit with the given

partition of variables

PSplit(n_parts::Int, model::JuMP.AbstractModel): Automatically partition

model variables into n_parts groups

Fields

partition::Vector{Vector{V}}: The partition of variables, where each inner

vector represents a group of variables that will be handled together

DisjunctiveProgramming.VariableProperties — Type

VariableProperties{L, U, F, S, SET, T}

A type for storing variable properties and attributes that can be applied to JuMP variables. This is used to capture and transfer variable information between models during reformulation.

Fields

info::JuMP.VariableInfo{L, U, F, S}: JuMP's VariableInfo struct containing bounds, fixed values, start values, and binary/integer constraints.

name::String: The variable name.

set::SET: The constraint set the variable belongs to (if any), obtained via JuMP.moi_set.

variable_type::T: The variable type information, critical for extensions.

Type Parameters

L, U, F, S: Type parameters from JuMP.VariableInfo for lower bound, upper bound, fixed value, and start value types.

SET: Type of the constraint set the variable belongs to.

T: Type of the variable type information.

Constructor VariableProperties(vref::JuMP.GenericVariableRef{T}) creates a VariableProperties instance from a JuMP variable reference, automatically extracting all relevant properties.

DisjunctiveProgramming.VariableProperties — Method

VariableProperties(expr)::VariableProperties

Creates a VariableProperties object with blank variable info (no bounds, not fixed, not binary/integer) from an expression. The expr argument is provided for extensions to infer additional properties.

Arguments

expr: Expression for extensions to extract metadata from

Returns

A VariableProperties object with blank info.

DisjunctiveProgramming._MOIAtLeast — Type

_MOIAtLeast <: AbstractCardinalitySet

MOI level set for AtLeast constraints, see AtLeast for recommended syntax.

DisjunctiveProgramming._MOIAtMost — Type

_MOIAtMost <: AbstractCardinalitySet

MOI level set for AtMost constraints, see AtMost for recommended syntax.

DisjunctiveProgramming._MOIExactly — Type

_MOIExactly <: AbstractCardinalitySet

MOI level set for Exactly constraints, see Exactly for recommended syntax.